+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * [GH-96] Ability to delete secret, delete secret versions and undelete a secret.
   * [GH-113] Extend the TOTP Secrets Engine with more APIs: Create Key, Read, Read All and Delete Key.
   * [GH-109] Support for Static Database Roles and Static Credentials. CRUD Role, create and rotate static credentials.
+  * Kerberos Auth Method: Support for Kerberos login tokens.
 
 **BREAKING CHANGES:**